+++ contracts/beacon/Voucher.sol
@@ -6,6 +6,7 @@ pragma solidity ^0.8.20;
 import {IexecLibOrders_v5} from "@iexec/poco/contracts/libs/IexecLibOrders_v5.sol";
 import {IexecPoco1} from "@iexec/poco/contracts/modules/interfaces/IexecPoco1.v8.sol";
 import {IexecPocoBoost} from "@iexec/poco/contracts/modules/interfaces/IexecPocoBoost.sol";
+import {IexecPocoAccessors} from "@iexec/poco/contracts/modules/interfaces/IexecPocoAccessors.sol";
 import {OwnableUpgradeable} from "@openzeppelin/contracts-upgradeable/access/OwnableUpgradeable.sol";
 import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import {IVoucherHub} from "../IVoucherHub.sol";
@@ -264,7 +265,15 @@ contract Voucher is OwnableUpgradeable, IVoucher {
         uint256 appPrice = appOrder.appprice;
         uint256 datasetPrice = datasetOrder.datasetprice;
         uint256 workerpoolPrice = workerpoolOrder.workerpoolprice;
-
+        uint256 volume = IexecPocoAccessors(iexecPoco).computeDealVolume(
+            appOrder,
+            datasetOrder,
+            workerpoolOrder,
+            requestOrder
+        );
+        uint256 dealPrice = datasetOrder.dataset != address(0)
+            ? (appPrice + datasetPrice + workerpoolPrice) * volume
+            : (appPrice + workerpoolPrice) * volume;
         sponsoredAmount = voucherHub.debitVoucher(
             voucherTypeId,
             appOrder.app,
@@ -272,11 +281,9 @@ contract Voucher is OwnableUpgradeable, IVoucher {
             datasetOrder.dataset,
             datasetPrice,
             workerpoolOrder.workerpool,
-            workerpoolPrice
+            workerpoolPrice,
+            volume
         );
-        // TODO: Compute volume and set dealPrice = taskPrice * volume instead of curent dealPrice
-        uint256 dealPrice = appPrice + datasetPrice + workerpoolPrice;
-
         if (sponsoredAmount != dealPrice) {
             // Transfer non-sponsored amount from the iExec account of the
             // requester to the iExec account of the voucher