+++ contracts/mocks/IexecPocoMock.sol
@@ -5,6 +5,7 @@ pragma solidity ^0.8.20;
 
 import {IexecLibOrders_v5} from "@iexec/poco/contracts/libs/IexecLibOrders_v5.sol";
 import {ERC20} from "@openzeppelin/contracts/token/ERC20/ERC20.sol";
+import {Math} from "@openzeppelin/contracts/utils/math/Math.sol";
 
 /**
  * @notice Testing purposes only.
@@ -21,15 +22,17 @@ contract IexecPocoMock is ERC20 {
         IexecLibOrders_v5.AppOrder calldata appOrder,
         IexecLibOrders_v5.DatasetOrder calldata datasetOrder,
         IexecLibOrders_v5.WorkerpoolOrder calldata workerpoolOrder,
-        IexecLibOrders_v5.RequestOrder calldata
+        IexecLibOrders_v5.RequestOrder calldata requestOrder
     ) external returns (bytes32 dealId) {
         if (shouldRevertOnSponsorMatchOrders) {
             revert("IexecPocoMock: Failed to sponsorMatchOrders");
         }
-        _burn(
-            msg.sender,
-            appOrder.appprice + datasetOrder.datasetprice + workerpoolOrder.workerpoolprice
-        );
+        uint256 volume = computeDealVolume(appOrder, datasetOrder, workerpoolOrder, requestOrder);
+
+        uint256 dealPrice = (appOrder.appprice +
+            datasetOrder.datasetprice +
+            workerpoolOrder.workerpoolprice) * volume;
+        _burn(msg.sender, dealPrice);
         return keccak256("deal");
     }
 
@@ -37,15 +40,16 @@ contract IexecPocoMock is ERC20 {
         IexecLibOrders_v5.AppOrder calldata appOrder,
         IexecLibOrders_v5.DatasetOrder calldata datasetOrder,
         IexecLibOrders_v5.WorkerpoolOrder calldata workerpoolOrder,
-        IexecLibOrders_v5.RequestOrder calldata
+        IexecLibOrders_v5.RequestOrder calldata requestOrder
     ) external returns (bytes32 dealId) {
         if (shouldRevertOnSponsorMatchOrdersBoost) {
             revert("IexecPocoMock: Failed to sponsorMatchOrdersBoost");
         }
-        _burn(
-            msg.sender,
-            appOrder.appprice + datasetOrder.datasetprice + workerpoolOrder.workerpoolprice
-        );
+        uint256 volume = computeDealVolume(appOrder, datasetOrder, workerpoolOrder, requestOrder);
+        uint256 dealPrice = (appOrder.appprice +
+            datasetOrder.datasetprice +
+            workerpoolOrder.workerpoolprice) * volume;
+        _burn(msg.sender, dealPrice);
         return keccak256("deal");
     }
 
@@ -54,5 +58,18 @@ contract IexecPocoMock is ERC20 {
     }
     function willRevertOnSponsorMatchOrdersBoost() external {
         shouldRevertOnSponsorMatchOrdersBoost = true;
+    }
+
+    function computeDealVolume(
+        IexecLibOrders_v5.AppOrder calldata appOrder,
+        IexecLibOrders_v5.DatasetOrder calldata datasetOrder,
+        IexecLibOrders_v5.WorkerpoolOrder calldata workerpoolOrder,
+        IexecLibOrders_v5.RequestOrder calldata requestOrder
+    ) public view returns (uint256) {
+        return
+            Math.min(
+                datasetOrder.dataset != address(0) ? datasetOrder.volume : type(uint256).max,
+                requestOrder.volume
+            );
     }
 }