+++ test/VoucherHub.test.ts
@@ -19,6 +19,7 @@ const duration = 3600;
 const voucherValue = 100;
 const asset = random();
 const assetPrice = 1;
+const volume = 3;
 const initVoucherHubBalance = 10 * voucherValue; // arbitrary value, but should support couple voucher creations
 
 describe('VoucherHub', function () {
@@ -41,6 +42,7 @@ describe('VoucherHub', function () {
             anyone,
         ] = await ethers.getSigners();
         const beacon = await voucherUtils.deployBeaconAndImplementation(admin.address);
+
         iexecPocoInstance = await new IexecPocoMock__factory()
             .connect(admin)
             .deploy()
@@ -681,10 +683,19 @@ describe('VoucherHub', function () {
         });
 
         it('Should debit voucher', async function () {
-            const sponsoredValue = BigInt(assetPrice * 3);
+            const sponsoredValue = BigInt(assetPrice * 3) * BigInt(volume);
             const voucherInitialCreditBalance = await voucherHub.balanceOf(voucher.address);
 
-            const args = [voucherType, asset, assetPrice, asset, assetPrice, asset, assetPrice] as [
+            const args = [
+                voucherType,
+                asset,
+                assetPrice,
+                asset,
+                assetPrice,
+                asset,
+                assetPrice,
+                volume,
+            ] as [
                 voucherTypeId: BigNumberish,
                 app: AddressLike,
                 appPrice: BigNumberish,
@@ -692,6 +703,7 @@ describe('VoucherHub', function () {
                 datasetPrice: BigNumberish,
                 workerpool: AddressLike,
                 workerpoolPrice: BigNumberish,
+                volume: BigNumberish,
             ];
             expect(await voucherHub.connect(voucher).debitVoucher.staticCall(...args)).to.be.equal(
                 sponsoredValue,
@@ -719,6 +731,7 @@ describe('VoucherHub', function () {
                         assetPrice,
                         asset,
                         assetPrice,
+                        volume,
                     );
             // The matcher 'emit' cannot be chained after or before 'reverted'
             // so we call several times to check different assertions
@@ -748,6 +761,7 @@ describe('VoucherHub', function () {
                         assetPrice,
                         asset,
                         assetPrice,
+                        volume,
                     ),
             ).to.not.emit(voucherHub, 'VoucherDebited');
             expect(await voucherHub.balanceOf(emptyVoucher.address))
@@ -770,6 +784,7 @@ describe('VoucherHub', function () {
                         assetPrice,
                         unEligibleAsset,
                         assetPrice,
+                        volume,
                     ),
             ).to.not.emit(voucherHub, 'VoucherDebited');
             expect(await voucherHub.balanceOf(voucher.address)).to.equal(initialCreditBalance);
@@ -781,7 +796,16 @@ describe('VoucherHub', function () {
             await expect(
                 voucherHub
                     .connect(anyone)
-                    .debitVoucher(999, asset, assetPrice, asset, assetPrice, asset, assetPrice),
+                    .debitVoucher(
+                        999,
+                        asset,
+                        assetPrice,
+                        asset,
+                        assetPrice,
+                        asset,
+                        assetPrice,
+                        volume,
+                    ),
             ).to.not.emit(voucherHub, 'VoucherDebited');
             expect(await voucherHub.balanceOf(anyone.address)).to.equal(initialCreditBalance);
         });